requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiogram-media-group==0.5.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.0.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.0.8
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.0
contourpy==1.2.0
cssutils==2.9.0
cycler==0.12.1
dataframe-image==0.2.2
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.1.0
fake-useragent==1.3.0
fastjsonschema==2.19.0
fiona==1.9.5
Flask==3.0.0
fonttools==4.45.0
fqdn==1.5.1
frozenlist==1.4.0
geographiclib==2.0
geopandas==0.14.1
geopy==2.4.1
h11==0.14.0
HeapDict==1.0.1
html2image==2.0.4.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-contrib-core==0.4.0
jupyter-contrib-nbextensions==0.5.1
jupyter-events==0.9.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-lsp==2.2.0
jupyter-nbextensions-configurator==0.5.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.1
kaleido==0.2.1
kiwisolver==1.4.5
lxml==4.9.3
magic-filter==1.0.12
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
multidict==6.0.4
mysql-connector-python==8.3.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.6
notebook==6.4.12
notebook_shim==0.2.3
numpy==1.26.2
openpyxl==3.1.2
outcome==1.2.0
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==4.0.0
plotly==5.18.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycountry==22.3.5
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.1.1
pyproj==3.6.1
pyrsistent==0.18.1
PySocks==1.7.1
pyTelegramBotAPI==4.16.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python2==1.2
pytz==2023.3.post1
PyYAML==6.0
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.1
scipy==1.11.4
seaborn==0.13.0
selenium==4.13.0
Send2Trash==1.8.2
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.5.1
telebot==0.0.5
tenacity==8.2.3
terminado==0.16.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
traitlets==5.13.0
trio==0.22.2
trio-websocket==0.10.4
types-python-dateutil==2.8.19.14
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
widgetsnbextension==4.0.9
workbook==1.1
wsproto==1.2.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
yapf==0.32.0
yarl==1.9.3
zipp==3.17.0
